% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_results.R
\name{filter_results}
\alias{filter_results}
\title{Filter results of multiHiCcompare}
\usage{
filter_results(hicexp, logfc_cutoff = 1, logcpm_cutoff = 1,
  p_cutoff = 0.01, alpha = 0.05, return_df = "pairedbed")
}
\arguments{
\item{hicexp}{A hicexp object which has been compared.}

\item{logfc_cutoff}{The logFC value you wish to filter
by. Defaults to 1.}

\item{logcpm_cutoff}{The logCPM cutoff you wish to
filter by. Defaults to 1.}

\item{p_cutoff}{The p-value cutoff you wish to filter
by. Defaults to 0.01.}

\item{alpha}{The p-value cut off for determining the count
of number of times a region is significant. Defaults
to 0.05.}

\item{return_df}{The format for the data.frame returned
by the function. Options are "bed" and "pairedbed".}
}
\value{
A data.table containing the filtered results.
}
\description{
Filter results of multiHiCcompare
}
\examples{
data('hicexp_diff')
filter_results(hicexp_diff)
}
