% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCT116_r5.R
\docType{data}
\name{HCT116_r5}
\alias{HCT116_r5}
\title{A 4 column sparse matrix for a Hi-C matrix.}
\format{
An object of class \code{data.frame} with 55092 rows and 4 columns.
}
\source{
Data from Rao 2017. See their website at
    \url{http://www.aidenlab.org/}
     Or the the GEO link to download the data
    \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE104888}
}
\usage{
HCT116_r5
}
\value{
A matrix
}
\description{
A matrix object with 4 columns and 55092 rows. 
    Sample 5 of 7 included as example Hi-C data.
    This is replicate 5 of HCT-116 cell
    line at 100KB resolution from chHCT116_r22.
}
\keyword{datasets}
