% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg19_cyto.R
\docType{data}
\name{hg19_cyto}
\alias{hg19_cyto}
\title{A GenomicRanges object containing centromeric, gvar, and stalk regions.}
\format{
An object of class \code{GRanges} of length 70.
}
\source{
Data from UCSC
    \url{http://hgdownload.cse.ucsc.edu/goldenPath/hg19/database/cytoBand.txt.gz}
}
\usage{
hg19_cyto
}
\value{
A GRanges object
}
\description{
A GRanges object with 2 metadata columns and 70 rows. 
    These ranges indicate the locations of centromeres,
    stalks, and gvar regions from hg19.
    Use this for filtering out these regions from 
    your data.
}
\keyword{datasets}
