% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg38_cyto.R
\docType{data}
\name{hg38_cyto}
\alias{hg38_cyto}
\title{A GenomicRanges object containing centromeric, gvar, and stalk regions.}
\format{An object of class \code{GRanges} of length 70.}
\source{
Data from UCSC
    \url{http://hgdownload.cse.ucsc.edu/goldenPath/hg38/database/cytoBand.txt.gz}
}
\usage{
hg38_cyto
}
\value{
A GRanges object
}
\description{
A GRanges object with 2 metadata columns and 70 rows. 
    These ranges indicate the locations of centromeres,
    stalks, and gvar regions from hg38.
    Use this for filtering out these regions from 
    your data.
}
\keyword{datasets}
