% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{hic_exactTest}
\alias{hic_exactTest}
\title{Perform exact test based difference detection on a Hi-C experiment}
\usage{
hic_exactTest(hicexp, parallel = FALSE, p.method = "fdr")
}
\description{
Perform exact test based difference detection on a Hi-C experiment
}
