% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{hic_glm}
\alias{hic_glm}
\title{Function to perform GLM differential analysis on Hi-C experiment}
\usage{
hic_glm(hicexp, design, parallel = FALSE)
}
\description{
Function to perform GLM differential analysis on Hi-C experiment
}
