% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_scaling.R
\name{hic_scale}
\alias{hic_scale}
\title{Perform library scaling on a hicexp object}
\usage{
hic_scale(hicexp)
}
\arguments{
\item{hicexp}{A hicexp object.}
}
\value{
A hicexp object.
}
\description{
Perform library scaling on a hicexp object
}
\details{
This function will perform library scaling 
    on a hicexp object. Scaling is performed 
    separately for each chromosome. This is an 
    alternative normalization method to the
    cyclic loess and fastlo methods also
    provided in multiHiCcompare. Use this normalization
    method if for some reason you do not want to remove
    trends in the data and only want to normalize based
    on library size.
}
\examples{
data("hicexp2")
hicexp2 <- hic_scale(hicexp2)

}
