% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{hic_table}
\alias{hic_table}
\alias{hic_table,Hicexp-method}
\alias{hic_table,character,ANY-method}
\title{Print the hic_table}
\usage{
hic_table(x)

\S4method{hic_table}{Hicexp}(x)
}
\arguments{
\item{x}{The Hicexp object}
}
\description{
Print the hic_table
}
