% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Hicexp-class}
\alias{Hicexp-class}
\title{An S4 class for working with Hi-C data}
\description{
An S4 class for working with Hi-C data
}
\section{Slots}{

\describe{
\item{\code{hic_table}}{A data.table containing the sparse upper triangular matrix 
for your Hi-C data.}

\item{\code{comparison}}{The results of a multiHiCcompare comparison.}

\item{\code{metadata}}{Data.frame for covariate information.}

\item{\code{resolution}}{The resolution of the dataset.}

\item{\code{normalized}}{Indicator for if data has been normalized.}
}}

