% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicexp2.R
\docType{data}
\name{hicexp2}
\alias{hicexp2}
\title{hicexp object with 4 samples from two groups.}
\format{
An object of class \code{Hicexp} of length 1.
}
\source{
Data from Rao 2017. See their website at
    \url{http://www.aidenlab.org/}
     Or the the GEO link to download the data
    \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE104888}
}
\usage{
hicexp2
}
\value{
A hicexp object
}
\description{
A hicexp object with a hic_table slot containing 666 rows
    from chromosome 22 at 1MB resolution.
}
\keyword{datasets}
