% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicexp_diff.R
\docType{data}
\name{hicexp_diff}
\alias{hicexp_diff}
\title{hicexp object with 7 samples from two groups.}
\format{An object of class \code{hicexp} of length 1.}
\source{
Data from Rao 2017. See their website at
    \url{http://www.aidenlab.org/}
     Or the the GEO link to download the data
    \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE104888}
}
\usage{
hicexp_diff
}
\value{
A hicexp object
}
\description{
A hicexp object with a hic_table slot containing 666 rows
    and a metadata slot containing 3 covariates. Same data
    as from "hicexp2" object but has been normalized and
    tested for differences with the hic_exactTest.
}
\keyword{datasets}
