% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{MD.composite}
\alias{MD.composite}
\title{Plot a composite MD plot with the results of a comparison}
\usage{
MD.composite(hicexp)
}
\description{
Plot a composite MD plot with the results of a comparison
}
