% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{MD.hicexp}
\alias{MD.hicexp}
\title{Make MD plots for all combinations of a condition}
\usage{
MD.hicexp(hicexp, prow = 3, pcol = 3)
}
\arguments{
\item{hicexp}{A hicexp object.}

\item{prow}{The number of rows to use for the 
grid of MD plots. Defaults to 3.}

\item{pcol}{The number of columns to use for
the grid of MD plots. Defaults to 3.}
}
\description{
Make MD plots for all combinations of a condition
}
