% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{MD_hicexp}
\alias{MD_hicexp}
\title{Make MD plots for all combinations of a condition}
\usage{
MD_hicexp(hicexp, prow = 3, pcol = 3, plot.chr = NA)
}
\arguments{
\item{hicexp}{A hicexp object.}

\item{prow}{The number of rows to use for the 
grid of MD plots. Defaults to 3.}

\item{pcol}{The number of columns to use for
the grid of MD plots. Defaults to 3.}

\item{plot.chr}{A specific chromosome or 
set of chromosome which you want to plot.
This should be a numeric value, i.e. to
plot chromosome 1 set plot.chr = 1, to
plot chromosomes 1 and 5 set plot.chr 
= c(1, 5). Defaults to NA indicating that
all chromosomes present in the hicexp
will be plotted.}
}
\value{
A set of MD plots.
}
\description{
Make MD plots for all combinations of a condition
}
\examples{
data("hicexp")
MD_hicexp(hicexp)
}
