% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,Hicexp-method}
\alias{meta,character,ANY-method}
\title{Print the metadata}
\usage{
meta(x)

\S4method{meta}{Hicexp}(x)
}
\arguments{
\item{x}{The Hicexp object}
}
\value{
Hicexp class information
}
\description{
Print the metadata
}
\examples{
data('hicexp')
meta(hicexp)
}
