% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot the count results from topDirs}
\usage{
plot_counts(dirs, plot.chr = NA)
}
\arguments{
\item{dirs}{The output of the topDirs function when the return_df 
option is set to "bed".}

\item{plot.chr}{A numeric value indicating a specific
chromosome number to subset the plot to. Defaults
to NA indicating that all chromosomes will be plotted.}
}
\value{
A plot.
}
\description{
Plot the count results from topDirs
}
\examples{
data('hicexp_diff')
dirs <- topDirs(hicexp_diff, return_df = 'bed')
plot_counts(dirs)
}
