% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r1.R
\docType{data}
\name{r1}
\alias{r1}
\title{A 4 column sparse matrix for a Hi-C matrix.}
\format{An object of class \code{data.frame} with 56603 rows and 4 columns.}
\source{
Data from Rao 2017. See their website at
    \url{http://www.aidenlab.org/}
     Or the the GEO link to download the data
    \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE104888}
}
\usage{
r1
}
\value{
A matrix
}
\description{
A matrix object with 4 columns and 56603 rows. 
    Sample 1 of 7 included as example Hi-C data.
}
\keyword{datasets}
