% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\docType{data}
\name{r6}
\alias{r6}
\title{A 4 column sparse matrix for a Hi-C matrix.}
\format{An object of class \code{data.frame} with 55581 rows and 4 columns.}
\source{
Data from Rao 2017. See their website at
    \url{http://www.aidenlab.org/}
     Or the the GEO link to download the data
    \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE104888}
}
\usage{
r6
}
\value{
A matrix
}
\description{
A matrix object with 4 columns and 55581 rows. 
    Sample 6 of 7 included as example Hi-C data.
}
\keyword{datasets}
