% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{resolution}
\alias{resolution}
\alias{resolution,Hicexp-method}
\alias{resolution,character,ANY-method}
\title{Print the resolution}
\usage{
resolution(x)

\S4method{resolution}{Hicexp}(x)
}
\arguments{
\item{x}{The Hicexp object}
}
\value{
Hicexp class information
}
\description{
Print the resolution
}
\examples{
data('hicexp')
resolution(hicexp)
}
