% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{results}
\alias{results}
\alias{results,Hicexp-method}
\alias{results,character,ANY-method}
\title{Print the results}
\usage{
results(x)

\S4method{results}{Hicexp}(x)
}
\arguments{
\item{x}{The Hicexp object}
}
\value{
Hicexp class information
}
\description{
Print the results
}
\examples{
data('hicexp')
results(hicexp)
}
