% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show,Hicexp-method}
\alias{show,Hicexp-method}
\alias{show,character,ANY-method}
\title{Print information about a HiCexp object}
\usage{
\S4method{show}{Hicexp}(object)
}
\arguments{
\item{object}{A Hicexp object}
}
\description{
Print information about a HiCexp object
}
