% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartApply.R
\name{smartApply}
\alias{smartApply}
\title{Function to apply either biocParallel or standard lapply}
\usage{
smartApply(parallel, x, FUN, ...)
}
\arguments{
\item{parallel}{Logical, should parallel processing be used?}

\item{x}{The main list object which the function will be applied to.}

\item{FUN}{The function to be applied.}

\item{...}{Additional arguments for bplapply or lapply.}
}
\value{
results of lapply or bplapply
}
\description{
Function to apply either biocParallel or standard lapply
}
