% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_mat.R
\docType{data}
\name{sparse_mat}
\alias{sparse_mat}
\title{A 3 column sparse matrix for a Hi-C matrix.}
\format{An object of class \code{data.frame} with 56603 rows and 3 columns.}
\source{
Data from Rao 2017. See their website at
    \url{http://www.aidenlab.org/}
     Or the the GEO link to download the data
    \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE104888}
}
\usage{
sparse_mat
}
\value{
A matrix
}
\description{
A matrix object with 3 columns and 56603 rows. 
    This is the same data as HCT116_r1, just
    without the chromosome column.
}
\keyword{datasets}
