% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastlo.R
\name{fastlo}
\alias{fastlo}
\title{Perform fast loess normalization on a Hi-C experiment}
\usage{
fastlo(hicexp, iterations = 3, span = 0.7, parallel = FALSE,
  verbose = TRUE, Plot = TRUE)
}
\arguments{
\item{hicexp}{A hicexp object}

\item{iterations}{The number of iterations (cycles) 
for fastlo to proceed through.}

\item{span}{The span of loess fitting. Defaults to 0.7. To
automatically calculate a span using the GCV set 
span = NA. However note that setting span = NA will
significantly slow down the normalization process.}

\item{parallel}{Logical. Should parallel processing be
used?}
}
\description{
Perform fast loess normalization on a Hi-C experiment
}
\details{
This function performs the fast loess (fastlo)
   normalization procedure on an hicexp object.
}
