% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{hic_glm}
\alias{hic_glm}
\title{Function to perform GLM differential analysis on Hi-C experiment}
\usage{
hic_glm(hicexp, design, contrast = NA, coef = NA, method = "QLFTest",
  M = 1, p.method = "fdr", parallel = FALSE, Plot = TRUE,
  max.pool = 0.7)
}
\arguments{
\item{hicexp}{A hicexp object,}

\item{design}{A design matrix for the GLM.}

\item{contrast}{A vector containing the contrast you want to test}

\item{coef}{A value or vect or the coefficients to test.}

\item{method}{The test method to be performed. Should be one of
"QLFTest", "LRTest", or "Treat".}

\item{M}{The log2 fold change value for a TREAT analysis.}

\item{parallel}{Logical, Should parallel processing be used?}

\item{Plot}{Logical, Should a composite MD plot be made for 
the results?}

\item{max.pool}{The proportion of unit distances after
which all further distances will be pooled. Distances
before this value will be progressively pooled and
any distances after this value will be combined
into a single pool. Defaults to 0.7. Warning: do
not adjust this value from the default unless you
are getting errors related to the lfproc function
or due to sparsity in fastlo normalization. If these
errors occur it is due to either sparsity or low 
variance and max.pool will need to be lowered; 
typically to 0.5 or 0.6.}
}
\description{
Function to perform GLM differential analysis on Hi-C experiment
}
\details{
This function performs the specified edgeR test
    on a per distance basis on the Hi-C data.
}
