% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hicexp.R
\name{make_hicexp}
\alias{make_hicexp}
\title{Make Hi-C experiment object from data}
\usage{
make_hicexp(..., groups, covariates = NULL)
}
\arguments{
\item{...}{Hi-C data. Data must in sparse upper triangular 
format with 4 columns: chr, region1, region2, IF.}

\item{groups}{A vector of the experimental groups 
corresponding to each Hi-C data object entered.}

\item{covariates}{Optional data.frame containing 
covariate information for your Hi-C experiment.
Some examples are enzyme used, batch number, etc.
Should have the same number of rows as the number
of Hi-C data objects entered and columns corresponding
to a covariates.}
}
\description{
Make Hi-C experiment object from data
}
\details{
Use this function to create a hicexp object for
    analysis in HiCcompare2.
}
