% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan_hicexp}
\alias{manhattan_hicexp}
\title{Manhattan plot function for results of multiHiCcompare}
\usage{
manhattan_hicexp(hicexp, method = "standard", return_df = FALSE)
}
\arguments{
\item{hicexp}{A hicexp object that has had differences
detected}

\item{method}{string denoting the p-value method to
use for plotting. Options are "standard", "fisher",
and "stouffer". "standard" plots a manhattan plot
using all individual p-values. "fisher" uses 
Fisher's method for combining p-values to combine
the p-values for each region which are then plotted.
"stouffer" uses the Stouffer-Liptak method for 
combining p-values for each region which are then 
plotted. "count" produces a plot where the y-values
are the -log10(1 / S) where S is the number of times
the region was found significant.}

\item{return_df}{Logical, should the data.frame used to
generate the plot be returned?}
}
\value{
A manhattan plot and optionally the data.frame used
    to generate the manhattan plot.
}
\description{
Manhattan plot function for results of multiHiCcompare
}
\details{
This function is used to create a manhattan
    plot for the significance of all genomic regions 
    in the dataset. These correspond to the rows (or columns)
    of the upper triangle of the full Hi-C matrix. Each genomic
    region of the Hi-C dataset has multiple interactions it
    is involved in and the significance of all of these can 
    be visualized with \code{method = "standard"}. 
    Alternatively the p-values for all these interactions
    can be combined using either Fisher's method or the
    Stouffer-Liptak method of combining p-values. Additionally
    the "count" option will plot based on the number of times
    each region was found to be involved in a signficantly 
    different interaction. The 
    manhattan plot can be used to identify "hotspot"
    regions of the genome where major differences
    seem to be located based on the results of a multiHiCcompare
    analysis.
}
\examples{
data("hicexp_diff")
manhattan_hicexp(hicexp_diff, method = "fisher")
}
