% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan_hicexp}
\alias{manhattan_hicexp}
\title{Manhattan plot function for results of HiCcompare2}
\usage{
manhattan_hicexp(hicexp, method = "standard", return_df = FALSE)
}
\arguments{
\item{hicexp}{A hicexp object that has had differences
detected}

\item{method}{string denoting the p-value method to
use for plotting. Options are "standard", "fisher",
and "stouffer". "standard" plots a manhattan plot
using all individual p-values. "fisher" uses 
Fisher's method for combining p-values to combine
the p-values for each region which are then plotted.
"stouffer" uses the Stouffer-Liptak method for 
combining p-values for each region which are then 
plotted.}

\item{return_df}{Logical, should the data.frame used to
generate the plot be returned?}
}
\description{
Manhattan plot function for results of HiCcompare2
}
