% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{MD.composite}
\alias{MD.composite}
\title{Plot a composite MD plot with the results of a comparison}
\usage{
MD.composite(hicexp, plot.chr = NA)
}
\arguments{
\item{hicexp}{A hicexp object which has 
had a multiHiCcompare comparison step performed on it.}

\item{plot.chr}{A specific chromosome or 
set of chromosome which you want to plot.
This should be a numeric value, i.e. to
plot chromosome 1 set plot.chr = 1, to
plot chromosomes 1 and 5 set plot.chr 
= c(1, 5). Defaults to NA indicating that
all chromosomes present in the hicexp
will be plotted.}
}
\value{
An MD plot
}
\description{
Plot a composite MD plot with the results of a comparison
}
\examples{
data("hicexp_diff")
MD.composite(hicexp_diff)
}
